Refactor Garak provider to support `module:` specifications for providers.
