+++ .gitignore
@@ -9,3 +9,4 @@ __pycache__/
 .DS_Store
 dist/
 .env
+_providers.d/
