--- providers.d/inline/eval/trustyai_garak.yaml
@@ -1,12 +1,0 @@
-module: llama_stack_provider_trustyai_garak
-config_class: llama_stack_provider_trustyai_garak.config.GarakEvalProviderConfig
-pip_packages:
-  - garak
-api_dependencies:
-  - inference
-  - files
-  - benchmarks
-optional_api_dependencies:
-  - safety
-  - shields
-  - telemetry
