+++ run-with-safety.yaml
@@ -19,6 +19,7 @@ providers:
   eval:
     - provider_id: trustyai_garak
       provider_type: inline::trustyai_garak
+      module: llama_stack_provider_trustyai_garak.inline
       config:
         base_url: ${env.BASE_URL:=http://localhost:8321/v1} # llama-stack service base url
         timeout: ${env.GARAK_TIMEOUT:=10800} # 3 hours max timeout for garak scan
@@ -92,4 +93,4 @@ shields:
 - shield_id: Prompt-Guard-86M
 server:
   port: 8321
-external_providers_dir: ./providers.d
+# external_providers_dir: ./providers.d
